Title: Implement brute force method to display bricks
Description:
Implemented a brute force method to display all the bricks on the game screen. Each brick is manually declared as an Actor object and positioned one after the other horizontally. And each new row of bricks is drawn in relation to the previous line.

Changes Made:
- Declared Actor objects for each type of brick (blue, green, grey, purple, red, yellow).
- Positioned each brick manually using absolute coordinates.
- Used while loops to draw bricks row by row until reaching the right edge of the screen.

Next Steps:
Refactor the code .